requirements.txt
Flask==1.1.2
Flask_SQLAlchemy==2.4.4
Flask-WTF==0.14.3
Flask-Migrate==2.6.0
Flask-Bootstrap4==4.0.2
Flask-FontAwesome==0.1.5
Flask-Login==0.5.0
Flask-Avatars==0.2.2
SQLAlchemy==1.3.23
python-dotenv==0.15.0
ansicolors==1.1.8
WTForms==2.3.3
Werkzeug==1.0.1
requests==2.25.1